\name{boostPM-package}
\alias{boostPM-package}
\alias{boostPM}
\docType{package}
\title{
\packageTitle{boostPM}
}
\description{
\packageDescription{boostPM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{boostPM}
\packageIndices{boostPM}
A package that runs the boosting algorithm introduced in Awaya and Ma (2021).
}
\author{
\packageAuthor{boostPM}

Maintainer: \packageMaintainer{boostPM}
  Naoki Awaya (naoki.awaya@duke.edu)
}
\references{
Awaya, N., & Ma, L. (2021). Tree boosting for learning probability measures. arXiv preprint arXiv:2101.11083.
}

