\name{boostPM-package}
\alias{boostPM-package}
\alias{boostPM}
\docType{package}
\title{
\packageTitle{boostPM}
}
\description{
\packageDescription{boostPM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{boostPM}
\packageIndices{boostPM}
}
\author{
\packageAuthor{boostPM}

Maintainer: \packageMaintainer{boostPM}
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
