\name{boostPM-package}
\alias{boostPM-package}
\alias{boostPM}
\docType{package}
\title{
\packageTitle{boostPM}
}
\description{
\packageDescription{boostPM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{boostPM}
\packageIndices{boostPM}
  An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{boostPM}

Maintainer: \packageMaintainer{boostPM}
}
\references{
  Literature or other references for background information
}
\keyword{ package }
\seealso{
  Optional links to other man pages, e.g.
  \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
  simple examples of the most important functions
}
